+++ src/deepblink/__init__.py
@@ -1,3 +1,17 @@
-"""deepBlink for spot detection and localization."""
+"""deepBlink for spot detection and localization.
 
-__version__ = "0.0.2"
+Modules are arranged as follows:
+- augment: Data augmentation to artificially increase dataset size.
+- cli: Command line interface for inferencing.
+- data: Data manipulation. Mainly to properly format for training.
+- datasets: Unique data import functions.
+- io: File-manipulation-related functions.
+- losses: Simple functions returning model losses.
+- metrics: Quantitative output of training / model performance.
+- models: Training loop containing classes for each type of model.
+- networks: Architecture / building of model structure.
+- optimizers: Simple functions returning model optimizers.
+- util: Basic utility functions not fitting into a category.
+"""
+
+__version__ = "0.0.3"