+++ src/deepblink/models/__init__.py
@@ -1,0 +1,7 @@
+"""Models folder.
+
+Contains the training loop and logic to handle data which feeds into the loop.
+"""
+
+from ._models import Model
+from .spots import SpotsModel