+++ docs/changelog.rst
@@ -2,6 +2,20 @@
 Changelog
 =========
 
+0.0.3 (2020-07-02)
+------------------
+
+* First custom code release.
+* All utility functions and training loop included.
+* CLI still cookiecutter and not functional.
+* First of five patches leading up to the first minor version:
+
+    * v0.0.3 Utility functions.
+    * v0.0.4 Command line interface for inference and script folder for training.
+    * v0.0.5 Evaluation added to measure performance on test dataset.
+    * v0.0.6 Addition of tests and completion of current TODO's.
+    * V0.0.7 Benchmarks/functions added to compare deepblink to other methods.
+
 0.0.2 (2020-06-29)
 ------------------